requirements.txt
anyio==4.4.0
certifi==2024.8.30
cffi==1.17.1
cryptography==41.0.1
exceptiongroup==1.2.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.8
peewee==3.16.2
pycparser==2.22
PyMySQL==1.1.0
python-decouple==3.8
python-telegram-bot==20.4
sniffio==1.3.1
typing_extensions==4.12.2
